% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metrics.R
\name{check_metrics}
\alias{check_metrics}
\title{Check Metrics}
\usage{
check_metrics(.metrics, .metric_data)
}
\arguments{
\item{.metrics}{Output from \code{\link[pcmh]{extract_metrics}}.}

\item{.metric_data}{Output from \code{\link[pcmh]{import_metric_data}}.}
}
\value{
Returns a \code{tibble} of \code{TRUE} and \code{FALSE} for each
metric check.
}
\description{
This function is a wrapper around \code{\link[pcmh]{.checks}}.
This function uses \code{\link[furrr]{future_pmap_dfr}} to map the
\code{\link[pcmh]{.checks}} function to each of the 26 metrics extraced
from \code{\link[pcmh]{extract_metrics}}.
}
\examples{
\dontrun{
check_metrics(.metrics, .metric_data)
}
}
