% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_report.R
\name{check_report}
\alias{check_report}
\title{Check Report}
\usage{
check_report(.path_pdf, .path_metric, .report_type = "PCMH")
}
\arguments{
\item{.path_pdf}{A \code{filepath} to the directory of the \code{.pdf} PCMH
reports.}

\item{.path_metric}{A \code{filepath} to the directory of the metric data and
descriptions pulled from oracle.}

\item{.report_type}{A \code{character string}; either "PCMH" or "Shared".}
}
\value{
Returns a \code{tibble} of metric checks for either the PCMH or
Shared Savings reports.
}
\description{
This function generates the output for the QA for either PCMH or Shared Savings
reports.
}
\examples{
\dontrun{
check_report(..path_pdf, .path_metric, .report_type = "PCMH")
}
}
