% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pdf_data.R
\name{clean_pdf_data}
\alias{clean_pdf_data}
\title{Clean pdf_data}
\usage{
clean_pdf_data(.list)
}
\arguments{
\item{.list}{A \code{list} that refers to a single PCMH report where each
\code{element} is a single page of the report.}
}
\value{
Returns \code{.list} as a single \code{tibble} with an added
\code{page} and \code{index} variable.
}
\description{
This function is meant to be used in conjunction with \code{\link[pcmh.qa]{read_pdf_data}}
to combine each page of the PCMH reports into a single \code{tibble} and
to add a \code{page} and \code{index} variable to the \code{tibble}.
}
