% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_index.R
\name{.add_index}
\alias{.add_index}
\title{Add Index}
\usage{
.add_index(.pdf_data)
}
\arguments{
\item{.pdf_data}{A \code{list} of \code{dataframes} read in with
\code{\link[pdftools]{pdf_data}}.}
}
\value{
Returns \code{.pdf_data} with an \code{index} column added to each
element.
}
\description{
Adds an \code{index} column to each of the elements of a \code{lsit}. Meant to
be used after you read in \code{.pdf} PCMH reports.
}
\examples{
\dontrun{
.add_index(list)
}
}
