% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_page_numbers.R
\name{.add_page_numbers}
\alias{.add_page_numbers}
\title{Add Page Numbers}
\usage{
.add_page_numbers(.pdf_data)
}
\arguments{
\item{.pdf_data}{A \code{list} of \code{dataframes} read in with
\code{\link[pdftools]{pdf_data}}.}
}
\value{
Returns \code{.pdf_data} with a \code{page} column added to each
element.
}
\description{
Adds a \code{page} column to each of the elements of a \code{lsit}. Meant to
be used after you read in \code{.pdf} PCMH reports.
}
\examples{
\dontrun{
.add_page_numbers(list)
}
}
