% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prvdr_num.R
\name{.add_prvdr_num}
\alias{.add_prvdr_num}
\title{Add prvdr_num}
\usage{
.add_prvdr_num(.data_frame)
}
\arguments{
\item{.data_frame}{A \code{dataframe}.}
}
\value{
Returns \code{.data_frame} as a \code{tibble} with an added
\code{prvdr_num} column.
}
\description{
This function adds a \code{prvdr_num} column to a \code{dataframe}. Helper
function meant to be used to clean the \code{.pdf} PCMH reports.
}
\examples{
\dontrun{
.add_prvdr_num(.data_frame)
}
}
