% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_true.R
\name{.all_true}
\alias{.all_true}
\title{All True}
\usage{
.all_true(.data_frame, .prvdr_num)
}
\arguments{
\item{.data_frame}{A \code{dataframe} read in by \code{\link[pcmh]{check_report}}.}

\item{.prvdr_num}{A Provider Number as a \code{numeric} value.}
}
\value{
Returns a tibble of \code{.prvdr_num} and either \code{TRUE} or
\code{FALSE}.
}
\description{
Helper function for \code{\link[pcmh]{.generate_summary_page}}.
}
