% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_care_categories.R
\name{.check_care_categories}
\alias{.check_care_categories}
\title{Check Care Categories}
\usage{
.check_care_categories(.care_categories, .care_categories_data)
}
\arguments{
\item{.care_categories}{A \code{dataframe} from \code{\link[pcmh]{.extract_care_categry}}}

\item{.care_categories_data}{A \code{dataframe} from \code{\link[pcmh]{import_cc_data}}.}
}
\value{
Returns a \code{tibble} of \code{TRUE}/\code{FALSE}.
}
\description{
This function checks the care category section from a PCMH report against the
data pulled from oracle.
}
\examples{
\dontrun{
.check_care_categories(.care_categories, .care_categories_data)
}
}
