% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.checks}
\alias{.checks}
\title{Checks}
\usage{
.checks(.metrics, .metric_data, .prvdr_num, .metric, .page)
}
\arguments{
\item{.metrics}{Output from \code{\link[pcmh]{extract_metrics}}.}

\item{.metric_data}{Output from \code{\link[pcmh]{import_metric_data}}.}

\item{.prvdr_num}{Specific provider number.}

\item{.metric}{Specific metric to be checked.}

\item{.page}{Specific page number the metric is on.}
}
\value{
Returns a \code{tibble} of \code{TRUE} if the value matched the data
from oracle or a \code{FALSE} if it did not.
}
\description{
This function is used to check a specific metric value from the \code{.pdf}
PCMH report agains the same metric value in the oracle data.
}
\examples{
\dontrun{
.checks(.metrics, .metric_data, .prvdr_num, .page, .metric)
}
}
