% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cc_data.R
\name{.clean_cc_data}
\alias{.clean_cc_data}
\title{Clean Care Category Data}
\usage{
.clean_cc_data(data_frame, n = 0)
}
\arguments{
\item{data_frame}{A \code{dataframe} of care category data.}

\item{n}{A positive \code{intiger} to be passed to \code{\link[pcmh]{round_df}}.}
}
\value{
Returns a \code{tibble} of clean, formatted care category data.
}
\description{
This function is part of \code{\link[pcmh]{import_cc_data}} and cleans and
formats the data for comparison.
}
\examples{
\dontrun{
.clean_cc_data(data_frame)
}
}
