% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metric_data.R
\name{.clean_metric_data}
\alias{.clean_metric_data}
\title{Clean metric_data}
\usage{
.clean_metric_data(.data_frame)
}
\arguments{
\item{.data_frame}{A \code{dataframe} read in by \code{\link[pcmh]{.read_metric_data}}.}
}
\value{
Returns a \code{.data_frame} as a \code{tibble} with clean variable
names.
}
\description{
This function takes the metric data read in by \code{\link[pcmh]{.read_metric_data}}
and uses \code{\link[janitor]{clean_names}} to clean the variable names.
}
\examples{
\dontrun{
.clean_metric_data(.data_frame)
}
}
