% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_care_category.R
\name{.extract_care_category}
\alias{.extract_care_category}
\title{Extract Care Category}
\usage{
.extract_care_category(.data_frame)
}
\arguments{
\item{.data_frame}{A \code{dataframe} read in by \code{\link[pcmh]{import_pdf_data}}.}
}
\value{
Returns a \code{tibble} of care category data from a \code{.pdf}
PCMH report.
}
\description{
This function uses \code{\link[furrr]{future_pmap_dfc}} to map
\code{\link[pcmh]{.filter_pdf_data}} to extract the columns of data from
the care category section of a \code{.pdf} PCMH report and arranges them
into a \code{tibble}.
}
\examples{
\dontrun{
.extract_care_categry(.data_frame)
}
}
