% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pdf_data.R
\name{.filter_pdf_data}
\alias{.filter_pdf_data}
\title{Filter pdf_data}
\usage{
.filter_pdf_data(.data_frame, .page, x_min, x_max, y_min, y_max)
}
\arguments{
\item{.data_frame}{\code{Dataframe} read in with \code{\link[pdftools]{pdf_data}}
that is to be filtered.}

\item{.page}{Page number.}

\item{x_min}{Lower bound of \code{x} range.}

\item{x_max}{Upper bound of \code{x} range.}

\item{y_min}{Lower bound of \code{y} range.}

\item{y_max}{Upper bound of \code{y} range.}
}
\value{
If text exists in the \code{x} and \code{y} range, returns the text
as a \code{tibble}, else returns \code{-99} as a \code{tibble}.
}
\description{
This is a helper function to be used in conjunction with \code{\link[pdftools]{pdf_data}}
and is used to extract the data from a \code{.pdf} file based on a certain
\code{x} and \code{y} range input.
}
\examples{
\dontrun{
.filter_pdf_data(df, 0, 50, 0, 150)
}
}
