% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_overview_page.R
\name{.generate_overview_page}
\alias{.generate_overview_page}
\title{Generate Overview Page}
\usage{
.generate_overview_page(.data_frame, .report_type = "PCMH")
}
\arguments{
\item{.data_frame}{Output of \code{\link[pcmh]{.generate_summary_page}}.}

\item{.report_type}{A \code{Character} string of either "PCMH" or "Shared Savings".}
}
\value{
Returns a 1 x 6 \code{tibble} of the overview statistics.
}
\description{
This function generates the overview statistics page of the output for the
PCMH QA.
}
\examples{
\dontrun{
.generate_overview_page(.data_frame, .report_type = "PCMH")
}
}
