% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary_page.R
\name{.generate_summary_page}
\alias{.generate_summary_page}
\title{Generate Summary Page}
\usage{
.generate_summary_page(.data_frame)
}
\arguments{
\item{.data_frame}{A \code{dataframe} read in by \code{\link[pcmh]{check_report}}.}
}
\value{
Returns an n x 2 \code{tibble} of \code{prvdr_num} and \code{TRUE} if
all the metrics are correct and \code{FALSE} if 1 or more are missing/incorrect.
}
\description{
This function generates the summary page for the PCMH Report QA output file.
}
\examples{
\dontrun{
.generate_summary_page(.data_frame)
}
}
