% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attribution_detail.R
\name{.get_attribution_detail}
\alias{.get_attribution_detail}
\title{Get Attribution Detail}
\usage{
.get_attribution_detail(.data_frame)
}
\arguments{
\item{.data_frame}{A cleaned \code{dataframe} referring to one PCMH report.}
}
\value{
Returns a \code{tibble} containing the information from the
Attribution Detail page of a PCMH report.
}
\description{
This function gets the information from the Attribution Detail page of the
PCMH report. Meant to be used after \code{\link[pcmh]{import_pdf_data}}.
}
\examples{
\dontrun{
.get_attribution_detial(.data_frame)
}
}
