% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metric.R
\name{.get_metric}
\alias{.get_metric}
\title{Get Metric}
\usage{
.get_metric(.data_frame, ...)
}
\arguments{
\item{.data_frame}{A \code{dataframe} you wish to extract data from.}

\item{...}{Extra parameters to be passed to \code{\link[pcmh]{.filter_pdf_data}}.}
}
\value{
Returns a \code{tibble} of the a specific metric.
}
\description{
This function is a wrapper around \code{\link[pcmh]{.filter_pdf_data}} and
extracts the specific \code{.metric} input.
}
\examples{
\dontrun{
.get_metric(.metric)
}
}
