% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metric_data.R
\name{.read_metric_data}
\alias{.read_metric_data}
\title{Read Metric Data}
\usage{
.read_metric_data(.path, ...)
}
\arguments{
\item{.path}{Filepath to the directory of metric data.}

\item{...}{Extra parameters to be passed to \code{\link[fs]{dir_ls}}.}
}
\value{
Returns a \code{tibble} of the \code{.csv} oracle data
}
\description{
This function uses \code{\link[vroom]{vroom}} to read in the \code{.csv}
oracle data to be checked against the \code{.pdf} PCMH report data.
}
\examples{
\dontrun{
.read_metric_data(.path)
}
}
