% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oracle_data.R
\name{.read_oracle_data}
\alias{.read_oracle_data}
\title{Read Oracle Data}
\usage{
.read_oracle_data(.path, ...)
}
\arguments{
\item{.path}{Filepath to the oracle data.}

\item{...}{Extra parameters to be passed to \code{\link[vroom]{vroom}}.}
}
\value{

}
\description{
This function reads in the \code{.csv} oracle data to be checked against the
\code{.pdf} PCMH report data.
}
