% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_pagination.R
\name{.remove_pagination}
\alias{.remove_pagination}
\title{Remove Pagination}
\usage{
.remove_pagination(.report, .n)
}
\arguments{
\item{.report}{A \code{list} of \code{dataframes} read in by
\code{\link[pcmh]{read_pdf_data}}.}

\item{.n}{A positive \code{intiger}.}
}
\value{
Returns a \code{list} of \code{dataframes}.
}
\description{
This function removes pages with from a PCMH report with less than \code{.n}
rows read in by \code{\link[pcmh]{read_pdf_data}} before it is cleaned.
}
\examples{
\dontrun{
.remove_pagination(.report)
}
}
