% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metrics2.R
\name{extract_metrics2}
\alias{extract_metrics2}
\title{Extract Metrics2}
\usage{
extract_metrics2(.data_frame, .met_desc)
}
\arguments{
\item{.data_frame}{A \code{dataframe} read in by \code{\link[pcmh]{import_pdf_data}}.}

\item{.met_desc}{A \code{dataframe} containing the metric descriptions read in
using \code{\link[pcmh]{.read_metric_data}}.}
}
\value{
Returns a \code{tibble} of metric data.
}
\description{
This function is an updated version of \code{\link[pcmh]{extract_metrics}} that
is compatible with the 2021 \code{PCMH} reports.
}
\examples{
\dontrun{
extract_metrics2(.data_frame)
}
}
