% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_care_category_page.R
\name{generate_care_category_page}
\alias{generate_care_category_page}
\title{Generate Care Category Page}
\usage{
generate_care_category_page(.care_categories, .care_categories_data)
}
\arguments{
\item{.care_categories}{A \code{dataframe} read in by.}

\item{.care_categories_data}{A \code{dataframe} read in by \code{\link[pcmh]{import_cc_data}}.}
}
\value{
Returns a \code{tibble}.
}
\description{
This fucntion maps \code{\link[pcmh]{.check_care_categories}} to generate an
output for PCMH QA.
}
\examples{
\dontrun{
generate_care_category_page(.care_categories, .care_categories_data)
}
}
