% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_detail.R
\name{generate_detail}
\alias{generate_detail}
\title{Generate Detail}
\usage{
generate_detail(.data_frame, .clean_data, .crosswalk)
}
\arguments{
\item{.data_frame}{A \code{dataframe} read in by \code{\link[pcmh]{import_pdf_data}}.}

\item{.clean_data}{Cleaned \code{dataframe} of metric data.}

\item{.crosswalk}{A \code{dataframe} of crosswalk for specific \code{PCMH} report type.}
}
\value{
Returns an \code{nx8} \code{dataframe} for \code{n} number of variables
extraced from a \code{PCMH} report.
}
\description{
This function uses inputs from a crosswalk to generate the \code{detail}
section of a \code{PCMH} QA report.
}
\examples{
\dontrun{
qa_report(data_frame, clean_data, crosswalk)
}
}
