% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ext.R
\name{get_ext}
\alias{get_ext}
\title{Get Extension}
\usage{
get_ext(.filepath)
}
\arguments{
\item{.filepath}{A \code{filepath}.}
}
\value{
Returns a \code{character string} of the file extension of \code{.filepath}.
}
\description{
This function takes a filepath as input and returns the extension as a
\code{character string}.
}
\examples{
\dontrun{
get_ext(.filepath)
}
}
