% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cc_data.R
\name{import_cc_data}
\alias{import_cc_data}
\title{Import Care Category Data}
\usage{
import_cc_data(path, ...)
}
\arguments{
\item{path}{\code{Filepath} to directory containing care category \code{.xlsx}.}

\item{...}{Extra parameters to be passed to \code{\link[pcmh]{.read_metric_data}}.}
}
\value{
Returns a \code{tibble}
}
\description{
Reads and cleans the Care Category data.
}
\examples{
\dontrun{
.import_cc_data(path, ...)
}
}
