% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pdf_data.R
\name{import_pdf_data}
\alias{import_pdf_data}
\title{Import pdf_data}
\usage{
import_pdf_data(.directory)
}
\arguments{
\item{.directory}{A path to a directory containing the \code{.pdf} PCMH
reports.}
}
\value{
Returns a \code{list} of \code{dataframes}. One for each \code{.pdf}
PCMH report in the directory.
}
\description{
This function is a wrapper around \code{\link[pcmh]{read_pdf_data}} and
\code{\link[pcmh]{clean_pdf_data}}.
This function reads in the \code{.pdf} PCMH reports and cleans the data
for future analysis.
}
\examples{
\dontrun{
import_pdf_data(~Documents)
}
}
