% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcmh_qa.R
\name{pcmh_qa}
\alias{pcmh_qa}
\title{PCMH QA}
\usage{
pcmh_qa(
  .pcmh_report_dir,
  .metric_data_dir,
  .metric_data_regexp,
  .report_type,
  .out_path = NA
)
}
\arguments{
\item{.pcmh_report_dir}{A \code{filepath} to the directory of PCMH
\code{.pdf} reports.}

\item{.metric_data_dir}{A \code{filepath} to the directory for metric data.}

\item{.metric_data_regexp}{A \code{regexp} for which metric data file to be
read in.}

\item{.report_type}{One of \code{"pcmh20"}, \code{"pool20"}, or
\code{"pcmh21"}.}

\item{.out_path}{A \code{filepath} to the location output will be saved.
Only supports \code{.xlsx} files and only writes if an input is supplied.
Defaults to \code{NA}.}
}
\value{
Retrurns a \code{list} of 2 \code{dataframes}.
}
\description{
This function is meant to complete an entire PCMH QA run in a single function.
}
\examples{
\dontrun{
pcmh_qa(.pcmh_report_dir,
  .metric_data_dir,
  .metric_data_regexp,
  .report_type,
  .out_path = NA)
}
}
