% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_report.R
\name{qa_report}
\alias{qa_report}
\title{Qa Report}
\usage{
qa_report(.path_pdf, .path_metric, .report_type = "PCMH")
}
\arguments{
\item{.path_pdf}{\code{Filepath} to the directory for reports.}

\item{.path_metric}{\code{Filepath} to the directory of metric data/descriptions.}

\item{.report_type}{A \code{character} string of either "PCMH" or "Shared".}
}
\value{
Returns a \code{list} of sheets for writing to excel.
}
\description{
This function generates the output for PCMH/Shared Savings report QA.
}
\examples{
\dontrun{
qa_report(.path_pdf, .path_metric, .report_type = "PCMH")
}
}
