% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Round2}
\usage{
round2(x, n)
}
\arguments{
\item{x}{A \code{numeric} value.}

\item{n}{An \code{integer} representing the decimal to round \code{x} to.}
}
\value{
Returns a \code{double}.
}
\description{
This function works the same as \code{\link[base]{round}} but always rounds
x.5 up and not to even.
}
\examples{
\dontrun{
round2(x, n)
}
}
