% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{Round Data Frame}
\usage{
round_df(data_frame, n = 0)
}
\arguments{
\item{data_frame}{A \code{dataframe} with \code{numeric} columns.}

\item{n}{A positive \code{intiger} to round to.}
}
\value{
Returns a \code{tibble} of with rounded \code{numeric} columns.
}
\description{
This function uses \code{\link[pcmh]{round2}} to round each of the \code{numeric}
columns in a dataframe.
}
\examples{
\dontrun{
round_df(data_frame, n = 0)
}
}
