% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Test Function}
\usage{
test_function(x)
}
\arguments{
\item{x}{input}
}
\value{
Returns \code{x}.
}
\description{
Placeholder for future function
}
\examples{
test_function("hi")
\dontrun{
test_function(hi)
}
}
