% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pdf_data.R
\name{clean_pdf_data}
\alias{clean_pdf_data}
\title{Clean pdf_data}
\usage{
clean_pdf_data(.list)
}
\arguments{
\item{.list}{A \code{list} of \code{dataframes} read in with
\code{\link[pcmh]{read_pdf_data}}.}
}
\value{
Returns \code{.list} as a single \code{tibble} with a \code{page} and
\code{index} column added to it.
}
\description{
This function takes a \code{list} of \code{dataframes} and adds an \code{index}
and \code{page} column then reduces them into a single \code{tibble}. This
function is a wrapper around \code{\link[pcmh]{.add_page_numbers}} and
\code{\link[pcmh]{.add_index}} and is meant to be used after you have read
in the PCMH reports using \code{\link[pcmh]{read_pdf_data}}.
}
\examples{
\dontrun{
clean_pdf_data(.list)
}
}
