% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_index.R
\name{.add_index}
\alias{.add_index}
\title{Add Index}
\usage{
.add_index(.data_frame)
}
\arguments{
\item{.data_frame}{A \code{dataframe}.}
}
\value{
Returns \code{.data_frame} as a \code{tibble} with an \code{index}
column added.
}
\description{
Adds an \code{index} column to a \code{dataframe}. Helper function to be used
to clean the \code{.pdf} PCMH reports data.
}
\examples{
\dontrun{
.add_index(dataframe)
}
}
