% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metric_data.R
\name{.clean_metric_data}
\alias{.clean_metric_data}
\title{Clean Metric Data}
\usage{
.clean_metric_data(.data_frame)
}
\arguments{
\item{.data_frame}{A \code{dataframe} read in by \code{\link[pcmh]{.read_metric_data}}.}
}
\value{
Returns a clean version of \code{.data_frame}.
}
\description{
This function takes the metric data read in by \code{\link[pcmh]{.read_metric_data}}
and formats the data to be passed to \code{\link[pcmh]{generate_detail}}.
}
\examples{
\dontrun{
.clean_metric_data(.data_frame)
}
}
