% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metric.R
\name{.get_metric}
\alias{.get_metric}
\title{Get Metric}
\usage{
.get_metric(.metric)
}
\arguments{
\item{.metric}{Input for which metric to extract.}
}
\value{
Returns a \code{tibble} of the specified \code{.metric}.
}
\description{
This function is a wrapper around \code{\link[pcmh]{.filter_pdf_data}} and
extracts the specific \code{.metric} input.
}
\examples{
\dontrun{
.get_metric(.metric)
}
}
