% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oracle_data.R
\name{.read_oracle_data}
\alias{.read_oracle_data}
\title{Read Oracle Data}
\usage{
.read_oracle_data(.path, ...)
}
\arguments{
\item{.path}{Filepath to the oracle data.}

\item{...}{Extra parameters to be passed to \code{\link[vroom]{vroom}}.}
}
\value{
Returns a \code{tibble} of the \code{.csv} oracle data
}
\description{
This function uses \code{\link[vroom]{vroom}} to read in the \code{.csv}
oracle data to be checked against the \code{.pdf} PCMH report data.
}
\examples{
\dontrun{
.read_oracle_data(.path)
}
}
