% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metrics.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Extract Metrics}
\usage{
extract_metrics(.data_frame)
}
\arguments{
\item{.data_frame}{A \code{dataframe} read in by \code{\link[pcmh]{import_pdf_data}}.}
}
\value{
Returns a \code{tibble} or metric data.
}
\description{
This function uses \code{\link[pcmh]{.get_metric}} function to extract each
metric from a \code{.pdf} PCMH report and arrange them into a \code{tibble}.
}
\examples{
\dontrun{
extract_metrics(.data_frame)
}
}
