% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metric_data.R
\name{import_metric_data}
\alias{import_metric_data}
\title{Import metric_data}
\usage{
import_metric_data(.path, .regexp = ".", ...)
}
\arguments{
\item{.path}{A \code{filepath} to the \code{.csv} metric data.}

\item{.regexp}{A \code{regular expression} to be passed to \code{\link[pcmh]{.read_metric_data}}}

\item{...}{Extra parameters to be passed to \code{\link[pcmh]{.read_metric_data}}.}
}
\value{
Returns \code{.path} as a \code{tibble} of clean data.
}
\description{
This function is a wrapper around \code{\link[pcmh]{.read_metric_data}} and
\code{\link[pcmh]{.clean_metric_data}}.
}
\examples{
\dontrun{
import_metric_data(.path)
}
}
