% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcmh_qa.R
\name{pcmh_qa}
\alias{pcmh_qa}
\title{PCMH QA}
\usage{
pcmh_qa(.path_pdf, .path_metric, .path_out)
}
\arguments{
\item{.path_pdf}{Path to the \code{directory} containing the \code{.pdf} PCMH
reports to be passed to \code{\link[pcmh]{import_pdf_data}}.}

\item{.path_metric}{Path to the \code{.csv} metric data to be passed to
\code{\link[pcmh]{import_metric_data}}.}

\item{.path_out}{Path to the location the output will be saved.}
}
\value{
Retrurns the character string "You completed PCMH QA!"
}
\description{
This function is a wrapper around \code{\link[pcmh]{import_pdf_data}} and
\code{\link[pcmh]{import_metric_data}}.
You call this function to generate output for the PCMH report QA.
}
\examples{
\dontrun{
pcmh_qa(.path_pdf, .path_metric, .path_out)
}
}
