% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdf_data.R
\name{read_pdf_data}
\alias{read_pdf_data}
\title{Test Function}
\usage{
read_pdf_data(path)
}
\arguments{
\item{path}{Filepath to the \code{.pdf} PCMH Reports}
}
\value{
Returns a \code{list} of \code{lists}. Each \code{element} of the
\code{list} refers to a PCMH Report and is a \code{list} of \code{dataframes},
each of which refers to a single page from the PCMH Report.
}
\description{
Placeholder for future function
}
\examples{
\dontrun{
read_pdf_data(hi)
}
}
